Starting to convert the sidebar into an overlay.
DECSLRM doesn't work in screen or tmux or on Shelly, and the
alternative method for implementing the sidebar (redraw the sidebar
when scrolling the swapbox) would be too inefficient. I'm planning
to convert the sidebar into an overlay, which would work similarly, but
it will be summoned (and dismissed) with a keypress and overlay the main
window. It'll do that by being a child of swappo.